P1/PR-3: extract web-research; one search entry, one cache
@launchstack/tools/web-research — the search execution machinery moves
down from trend-search (providers registry, executeSearch with retry +
URL dedup + strategy selection, Exa/Serper clients); the trend-search
vertical keeps query planning, synthesis, jobs, and Inngest
orchestration and imports the tool. Env reads (EXA_API_KEY,
SERPER_API_KEY, SEARCH_PROVIDER) now live only in the tool's config.ts,
per the tools lint rule.

- createTtlCache: the 45-line sha256+TTL+prune cache that existed twice
  (trend-search 1h/100, competitor 24h/50) is one mechanism; each caller
  keeps its own key shape and constants.
- marketing competitor.ts stops deep-importing trend-search internals
  and drops its private cache — both search consumers now use the tool.
- Old import paths keep working via re-export shims (web-search.ts,
  providers/*, types) — the four web test suites that import them
  directly pass unchanged.

Verified: tools typecheck + 23 vitest tests green; features and web
typecheck clean; full web Jest suite green (174 suites / 2107 tests).
