+++ packages/features/src/trend-search/cache.ts
@@ -2,45 +2,26 @@
  * In-memory cache for trend search results.
  * Reduces redundant API calls (Exa/Serper) and LLM work when the same
  * query + company context is used within the TTL window.
+ *
+ * Mechanism lives in @launchstack/tools/web-research (`createTtlCache`,
+ * unification PR-3); this module owns the trend-search key shape and TTL.
  */
-import { createHash } from "node:crypto";
+import { createTtlCache } from "@launchstack/tools/web-research";
 import type { TrendSearchOutput } from "./types";
 
 const TTL_MS = 60 * 60 * 1000; // 1 hour – trends don't change that fast
 
-interface CacheEntry {
-    output: TrendSearchOutput;
-    expiresAt: number;
-}
-
-const cache = new Map<string, CacheEntry>();
+const cache = createTtlCache<TrendSearchOutput>({ ttlMs: TTL_MS, maxEntries: 100 });
 
 function buildCacheKey(query: string, companyContext: string): string {
-    const normalized = `${query.trim().replace(/\s+/g, " ")}::${companyContext.trim().replace(/\s+/g, " ")}`;
-    return createHash("sha256").update(normalized).digest("hex");
-}
-
-function prune(): void {
-    const now = Date.now();
-    for (const [key, entry] of cache.entries()) {
-        if (entry.expiresAt <= now) {
-            cache.delete(key);
-        }
-    }
+    return `${query.trim().replace(/\s+/g, " ")}::${companyContext.trim().replace(/\s+/g, " ")}`;
 }
 
 export function getCachedTrendSearch(
     query: string,
     companyContext: string
 ): TrendSearchOutput | null {
-    const key = buildCacheKey(query, companyContext);
-    const entry = cache.get(key);
-    if (!entry) return null;
-    if (entry.expiresAt <= Date.now()) {
-        cache.delete(key);
-        return null;
-    }
-    return entry.output;
+    return cache.get(buildCacheKey(query, companyContext));
 }
 
 export function setCachedTrendSearch(
@@ -48,12 +29,5 @@ export function setCachedTrendSearch(
     companyContext: string,
     output: TrendSearchOutput
 ): void {
-    if (cache.size > 100) {
-        prune();
-    }
-    const key = buildCacheKey(query, companyContext);
-    cache.set(key, {
-        output,
-        expiresAt: Date.now() + TTL_MS,
-    });
+    cache.set(buildCacheKey(query, companyContext), output);
 }
