+++ packages/tools/README.md
@@ -17,15 +17,16 @@ contract types (`ToolRunContext`, `ToolResult`, `ToolError`) live in
 
 ## Catalog
 
-| Tool                   | Import                                  | What it does                                                                                                                                                                                                              | Consumers                                                             |
-| ---------------------- | --------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | --------------------------------------------------------------------- |
-| **company-context**    | `@launchstack/tools/company-context`    | Company identity (one DB read), knowledge-base context assembly, and CompanyDNA extraction (metadata-first with RAG fallback). Owns the `company_metadata` data contract (`./schema`) and the `readFact` confidence gate. | marketing-pipeline, email-pipeline, company-metadata (schema)         |
-| **grounded-retrieval** | `@launchstack/tools/grounded-retrieval` | Company-scoped RAG retrieval with named snippet policies (topK / weights / caps) and a declared empty-vs-throw failure policy.                                                                                            | company-context, brand-voice, persona; claim-evidence (planned, PR-4) |
-| **brand-voice**        | `@launchstack/tools/brand-voice`        | Synthesize a BrandVoice profile (tone, vocabulary, sentence style, formality) from the company's own documents.                                                                                                           | marketing-pipeline; email tone rules (planned)                        |
-| **persona**            | `@launchstack/tools/persona`            | Synthesize a TargetPersona (role, pain points, priorities, language style) for a described audience, grounded in company knowledge.                                                                                       | marketing-pipeline                                                    |
+| Tool                   | Import                                  | What it does                                                                                                                                                                                                              | Consumers                                                                          |
+| ---------------------- | --------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ---------------------------------------------------------------------------------- |
+| **company-context**    | `@launchstack/tools/company-context`    | Company identity (one DB read), knowledge-base context assembly, and CompanyDNA extraction (metadata-first with RAG fallback). Owns the `company_metadata` data contract (`./schema`) and the `readFact` confidence gate. | marketing-pipeline, email-pipeline, company-metadata (schema)                      |
+| **grounded-retrieval** | `@launchstack/tools/grounded-retrieval` | Company-scoped RAG retrieval with named snippet policies (topK / weights / caps) and a declared empty-vs-throw failure policy.                                                                                            | company-context, brand-voice, persona; claim-evidence (planned, PR-4)              |
+| **brand-voice**        | `@launchstack/tools/brand-voice`        | Synthesize a BrandVoice profile (tone, vocabulary, sentence style, formality) from the company's own documents.                                                                                                           | marketing-pipeline; email tone rules (planned)                                     |
+| **persona**            | `@launchstack/tools/persona`            | Synthesize a TargetPersona (role, pain points, priorities, language style) for a described audience, grounded in company knowledge.                                                                                       | marketing-pipeline                                                                 |
+| **web-research**       | `@launchstack/tools/web-research`       | Provider-pluggable web search (Exa/Serper registry, strategy selection, retry, URL dedup) plus the shared in-memory TTL cache (`createTtlCache`). Env reads live in its `config.ts`.                                      | trend-search (planning/synthesis/jobs stay there), marketing competitor + research |
 
-Planned (design doc, Phase 1): `web-research`, `claim-evidence`,
-`content-scoring`, `social-publish`, `platform-profiles`.
+Planned (design doc, Phase 1): `claim-evidence`, `content-scoring`,
+`social-publish`, `platform-profiles`.
 
 ## Tests
