+++ packages/tools/src/web-research/providers/registry.ts
@@ -1,0 +1,8 @@
+import type { SearchProviderFn } from "../types";
+import { callExa } from "./exa";
+import { callSerper } from "./serper";
+
+export const providerRegistry: Record<string, SearchProviderFn> = {
+    exa: callExa,
+    serper: callSerper,
+};
